% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnm-block.R
\name{gnm_block}
\alias{gnm_block}
\title{Block a feature in the network}
\usage{
gnm_block(net, gfid)
}
\arguments{
\item{net}{A \link{gnm_network} object (must be open).}

\item{gfid}{Global Feature ID to block.}
}
\value{
\code{net} invisibly (for piping).
}
\description{
Blocked features are excluded from path calculations, simulating
e.g. a broken pipe or closed road.
}
