% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnm-block.R
\name{gnm_block_all}
\alias{gnm_block_all}
\title{Block all features in the network}
\usage{
gnm_block_all(net)
}
\arguments{
\item{net}{A \link{gnm_network} object (must be open).}
}
\value{
\code{net} invisibly (for piping).
}
\description{
Block all features in the network
}
