% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R, R/gnm-analysis.R
\name{gnm_components}
\alias{gnm_components}
\title{GNM Connected Components Result}
\usage{
gnm_components(net)

gnm_components(net)
}
\arguments{
\item{net}{A \link{gnm_network} object (must be open).}
}
\value{
A \link{gnm_components} object.
}
\description{
S7 class holding connected component analysis results.

Identifies groups of features that are connected to each other
in the network graph.
}
