% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnm-rules.R
\name{gnm_delete_all_rules}
\alias{gnm_delete_all_rules}
\title{Delete all rules from the network}
\usage{
gnm_delete_all_rules(net)
}
\arguments{
\item{net}{A \link{gnm_network} object (must be open).}
}
\value{
\code{net} invisibly (for piping).
}
\description{
Delete all rules from the network
}
