% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnm-rules.R
\name{gnm_delete_rule}
\alias{gnm_delete_rule}
\title{Delete a specific rule from the network}
\usage{
gnm_delete_rule(net, rule)
}
\arguments{
\item{net}{A \link{gnm_network} object (must be open).}

\item{rule}{Rule string to delete.}
}
\value{
\code{net} invisibly (for piping).
}
\description{
Delete a specific rule from the network
}
