% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnm-analysis.R
\name{gnm_k_paths}
\alias{gnm_k_paths}
\title{Find K shortest paths between two features}
\usage{
gnm_k_paths(net, from, to, k = 3L)
}
\arguments{
\item{net}{A \link{gnm_network} object (must be open).}

\item{from}{Source feature GFID.}

\item{to}{Target feature GFID.}

\item{k}{Number of paths to find (default 3).}
}
\value{
A \link{gnm_path} object. The features data.frame contains all K paths;
individual paths can be identified by breaks in the sequence.
}
\description{
Uses Yen's algorithm (which internally uses Dijkstra) to find the
K shortest paths between two features.
}
