% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnm-rules.R
\name{gnm_rules}
\alias{gnm_rules}
\title{List network rules}
\usage{
gnm_rules(net)
}
\arguments{
\item{net}{A \link{gnm_network} object (must be open).}
}
\value{
Character vector of rule strings.
}
\description{
List network rules
}
