% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnm-block.R
\name{gnm_unblock}
\alias{gnm_unblock}
\title{Unblock a feature in the network}
\usage{
gnm_unblock(net, gfid)
}
\arguments{
\item{net}{A \link{gnm_network} object (must be open).}

\item{gfid}{Global Feature ID to unblock.}
}
\value{
\code{net} invisibly (for piping).
}
\description{
Unblock a feature in the network
}
