% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnm-block.R
\name{gnm_unblock_all}
\alias{gnm_unblock_all}
\title{Unblock all features in the network}
\usage{
gnm_unblock_all(net)
}
\arguments{
\item{net}{A \link{gnm_network} object (must be open).}
}
\value{
\code{net} invisibly (for piping).
}
\description{
Unblock all features in the network
}
