% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{osgeo0-package}
\alias{osgeo0}
\alias{osgeo0-package}
\title{osgeo0: R Interface to GDAL Geographic Network Model}
\description{
Idiomatic R bindings for the GDAL Geographic Network Model (GNM) via
the Python osgeo.gnm module and reticulate. Create, manage, and analyse
geographic networks with topology building, shortest-path analysis,
and connected component detection.

Uses \link[wk:wk-package]{wk::wk} for geometry interchange and \link{S7} for the class system.
}
\section{Getting started}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{net <- gnm_create(tempdir(), "my_network", srs = "EPSG:4326")
gnm_import(net, wells_df, "wells")
gnm_import(net, pipes_df, "pipes")
gnm_connect_auto(net, tolerance = 0.000001)
path <- gnm_shortest_path(net, from = 1, to = 10)
gnm_close(net)
}\if{html}{\out{</div>}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/osgeo0}
  \item Report bugs at \url{https://github.com/mdsumner/osgeo0/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
