% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-wk.R
\name{wk_handle.gnm_components}
\alias{wk_handle.gnm_components}
\title{Handle wk operations on gnm_components objects}
\usage{
wk_handle.gnm_components(handleable, handler, ...)
}
\arguments{
\item{handleable}{A \link{gnm_components} object.}

\item{handler}{A wk handler.}

\item{...}{Passed to \code{\link[wk:wk_handle]{wk::wk_handle()}}.}
}
\value{
Result of the handler.
}
\description{
Handle wk operations on gnm_components objects
}
