% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-wk.R
\name{wk_handle.gnm_path}
\alias{wk_handle.gnm_path}
\title{Handle wk operations on gnm_path objects}
\usage{
wk_handle.gnm_path(handleable, handler, ...)
}
\arguments{
\item{handleable}{A \link{gnm_path} object.}

\item{handler}{A wk handler.}

\item{...}{Passed to \code{\link[wk:wk_handle]{wk::wk_handle()}}.}
}
\value{
Result of the handler.
}
\description{
Delegates to the underlying wk geometry column in the features data.frame.
}
